requirements.txt
# LLM
openai>=1.30.0

# Embedding & Vector
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4

# Document Parsing
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-pptx>=0.6.23

# Web UI
streamlit>=1.37.0
streamlit-agraph>=0.0.45

# REST API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.18
sse-starlette>=2.0.0

# Utils
python-dotenv>=1.0.0
